requirements.txt
openai==0.10.2
python-dotenv==0.19.1
